+++ src/gocept/pseudonymize/__init__.py
@@ -13,6 +13,7 @@ from gocept.pseudonymize.pseudonymize import (
     name,
     phone,
     street,
+    string,
     text,
     time,
     year,